refactor[next]: address review on test_basic split
- new test_broadcast.py: the explicit broadcast() tests (moved from
  test_gt4py_builtins.py) plus the implicit-broadcast / zero-dim tests
  (moved from test_basic.py)
- drop test_undefined_symbols from test_basic.py: it duplicates the
  existing unit test in unit_tests/ffront_tests/test_type_deduction.py
  (same 'Undeclared symbol' DSLError check), so it no longer needs to
  fan out over the backend matrix
- test_basic.py now holds only trivial-execution smoke tests
  (copy, infinity, nan, docstring)
